==> limpieza_y_extraccion_de_datos/5_asignar_categoria_startups.ipynb <==
import pandas as pd

# Leer el archivo con separador ;
df = pd.read_csv("datasets/categorias_startups.csv", sep=",")

df.head()

# Función que asigna la primera categoría principal reconocida
def asignar_categoria_prioritaria(categorias_str):
    if pd.isna(categorias_str):
        return "No Especificado"
    categorias = [c.strip() for c in categorias_str.split(",")]
    categorias_mapeadas = [mapeo_categorias.get(c, "No Especificado") for c in categorias]
    for categoria in prioridad:
        if categoria in categorias_mapeadas:
            return categoria
    return "No Especificado"

# Supongamos que la columna se llama 'categories' y tiene valores separados por coma
df["categories"] = df["categories"].fillna("Unspecified")

# Separar por coma y hacer una fila por cada categoría
df_exploded = df.assign(category_split=df["categories"].str.split(",")).explode("category_split")

# Quitar espacios
df_exploded["category_split"] = df_exploded["category_split"].str.strip()

# Diccionario de mapeo a las 9 categorías principales
# 2. Mapeo de subcategorías a categorías principales (en español)
mapeo_categorias = {
    "Real Estate": "Bienes Raíces y Construcción",
    "Construction": "Bienes Raíces y Construcción",
    "Education": "Educación",
    "Financial Services": "Finanzas",
    "Lending and Investments": "Finanzas",
    "FinTech": "Finanzas",
    "Health Care": "Salud",
    "Biotech": "Salud",
    "Biotechnology": "Salud",
    "Industrial": "Industria",
    "Manufacturing": "Industria",
    "B2B": "Software y Servicios B2B",
    "SaaS": "Software y Servicios B2B",
    "Software": "Software y Servicios B2B",
    "Artificial Intelligence (AI)": "Software y Servicios B2B",
    "Data and Analytics": "Software y Servicios B2B",
    "Hardware": "Software y Servicios B2B",
    "Information Technology": "Software y Servicios B2B",
    "Consumer Goods": "Consumo",
    "Food and Beverage": "Consumo",
    "Retail": "Consumo",
    "Commerce and Shopping": "Consumo",
    ##"Government": "Gobierno",
     #"Transportation": "Transporte",
     #"Agriculture and Farming": "Agricultura",
     #"Administrative Services": "Servicios Profesionales",
     #"Professional Services": "Servicios Profesionales"
}

prioridad = [
    "Bienes Raíces y Construcción",
    "Industria",
    "Educación",
    "Finanzas",
    "Software y Servicios B2B",
    #"Gobierno",
    "Salud",
    "Consumo",
    #"Transporte",
    #"Agricultura",
    #"Servicios Profesionales",
    "No Especificado"
]


# Aplicar función a cada fila
df["main_category"] = df["categories"].apply(asignar_categoria_prioritaria)

# Guardar resultado
df.to_csv("datasets/categorias_startups_transformadas.csv", index=False)
