==> limpieza_y_extraccion_de_datos/6_tareas_de_limpieza_datos.ipynb <==
import json
import pandas as pd
from datetime import datetime
from datetime import timedelta

# Leer el archivo CSV con separador coma
df = pd.read_csv("datasets/startups_etapa_temprana.csv", sep=",")
df.head()

len(df)

# Leer el archivo con separador ;
fundadores_df = pd.read_csv("datasets/conteo_fundadores_por_startup.csv", sep=";")
fundadores_df.head()

# Hacer el merge (cruce) por company_name
df = df.merge(
    fundadores_df[["company_name", "num_female_founders", "num_male_founders"]],
    on="company_name",
    how="left"
)

len(df)

# Crear columnas binarias sin sobrescribir las originales
df["has_linkedin"] = df["linkedin"].apply(lambda x: 1 if pd.notna(x) and str(x).strip() != "" else 0)
df["has_facebook"] = df["facebook"].apply(lambda x: 1 if pd.notna(x) and str(x).strip() != "" else 0)

len(df)

# Si no tiene LinkedIn, colocar 0 en followers y empleados
df["linkedin_followers"] = df["has_linkedin"].apply(lambda x: 0 if not x else pd.NA)
df["linkedin_employees"] = df["has_linkedin"].apply(lambda x: 0 if not x else pd.NA)

# Si no tiene Facebook, colocar 0 en followers
df["facebook_followers"] = df["has_facebook"].apply(lambda x: 0 if not x else pd.NA)


# Reordenar columnas para insertar después de 'linkedin'
col_list = list(df.columns)
idx = col_list.index("linkedin")

# Insertar las nuevas columnas en la posición deseada
for col_name in ["linkedin_employees", "linkedin_followers"][::-1]:  # insert en orden inverso
    col_list.insert(idx + 1, col_list.pop(col_list.index(col_name)))

# Aplicar nuevo orden
df = df[col_list]

# Reordenar columnas para insertar después de 'linkedin'
col_list = list(df.columns)
idx = col_list.index("facebook")

# Insertar las nuevas columnas en la posición deseada
for col_name in ["facebook_followers"][::-1]:  # insert en orden inverso
    col_list.insert(idx + 1, col_list.pop(col_list.index(col_name)))

# Aplicar nuevo orden
df = df[col_list]

# Leer el archivo con separador ;
df_x = pd.read_csv("datasets/cuentas_startup_twitter_con_seguidores.csv", sep=";")
df_x.head()

len(df_x)

#duplicados = df[df["company_name"].duplicated()]
#print(duplicados["company_name"])

# Hacer el merge temporalmente solo para obtener la columna de twitter
twitter_merge = pd.merge(df[["company_name"]], df_x[["company_name", "twitter"]], on="company_name", how="left")

# Crear nueva columna has_twitter directamente en df
df["has_twitter"] = twitter_merge["twitter"].apply(lambda x: 1 if pd.notna(x) and str(x).strip() != "" else 0)

# Verificamos el resultado
print(df[["company_name", "has_twitter"]].head())

len(df)

# Hacemos merge por 'name' para traer los followers
df = pd.merge(df, df_x[["company_name", "twitter_followers"]], on="company_name", how="left")

# Rellenar con 0 donde no haya seguidores
df["twitter_followers"] = pd.to_numeric(df["twitter_followers"], errors="coerce").fillna(0).astype(int)

# Verificamos el resultado
print(df[["company_name", "twitter_followers"]].head())

len(df)

df = df.drop(["twitter", "founders"], axis=1)

columnas = ["num_articles", "num_investments_funding_rounds", "ipqwery_num_trademark_registered","ipqwery_num_patent_granted",] 

df[columnas] = df[columnas].replace("", pd.NA)  # convierte "" en NaN
df[columnas] = df[columnas].fillna(0)           # reemplaza NaN por 0


df[columnas] = df[columnas].astype("Int64")

len(df)

# Leer el archivo CSV con separador coma
df_c = pd.read_csv("datasets/categorias_startups_transformadas.csv", sep=",")
df_c.head()

# Hacer el merge temporalmente solo para obtener la columna de twitter
main_category = pd.merge(df[["company_name"]], df_c[["company_name", "main_category"]], on="company_name", how="left")

# Crear nueva columna has_twitter directamente en df
df["category"] = main_category["main_category"]

# Verificamos el resultado
print(df[["company_name", "category"]].head())

df.head()

df = df.drop(["categories"], axis=1)
len(df)

# Se trabajará solo con startups activas
df = df[df["operating_status"] != "closed"]
len(df)

#Calcular edad de la compañia

# Asegurar que la columna founded_on esté en formato fecha
df["founded_on"] = pd.to_datetime(df["founded_on"], errors="coerce")

# Fecha de corte para el cálculo
hoy = pd.to_datetime("today")

#fecha ficticia para compañias con estado closed
fecha_cierre_ficticia = pd.to_datetime("2020-12-31")

df["company_age"] = df.apply(
    lambda row: ((fecha_cierre_ficticia if row["operating_status"] == "closed" else hoy) - row["founded_on"]).days // 365
    if pd.notna(row["founded_on"]) else pd.NA,
    axis=1
).astype("Int64")


len(df)

#Se trabajará con startups fundadas entre el 2014 y 2019
df = df[
    (df["founded_on"].dt.year >= 2014) &
    (df["founded_on"].dt.year <= 2019)
]
len(df)

# Guardar 
df.to_csv("datasets/startups_etapa_temprana_final.csv", index=False)

# Guardar 

df_procesamiento_manual =  df[["company_name","link_company", "linkedin", "linkedin_employees", "linkedin_followers", "facebook", "facebook_followers"]]

df_procesamiento_manual.to_csv("datasets/startups_etapa_temprana_para_procesamiento_manual.csv", index=False)

len(df)
