==> limpieza_y_extraccion_de_datos/7_limpieza_final_y_generacion_del_conjunto_datos.ipynb <==
import json
import pandas as pd

# Leer el archivo CSV con separador coma
df = pd.read_csv("datasets/startups_etapa_temprana_final.csv", sep=",")
df.head()

# Leer el archivo CSV con separador coma
df_startups_redes_sociales = pd.read_csv("datasets/startups_etapa_temprana_datos_redes_sociales.csv", sep=";")
df_startups_redes_sociales.head()


len(df_startups_redes_sociales)

df = df.drop(columns=["linkedin_employees", "linkedin_followers", "facebook_followers"], errors="ignore")

df = df.merge(
    df_startups_redes_sociales[["company_name", "linkedin_employees", "linkedin_followers", "facebook_followers", "selected"]],
    on="company_name",
    how="left"
)

len(df)

df = df[df_startups_redes_sociales["selected"] == 1]
len(df)

df = df.drop(["linkedin","facebook", "selected", "link_company","num_employees_enum", "has_linkedin", "has_facebook", "has_twitter"], axis=1)

df["years_since_last_funding"] = (pd.to_datetime("2025-07-10") - pd.to_datetime(df["last_funding_at"])).dt.days / 365

# Diccionario de traduccion sin tildes
column_translation = {
    
     # Datos de la empresa
    "company_name": "nombre_de_la_empresa",
    "founded_on": "fecha_de_fundacion",
    "ipqwery_num_trademark_registered": "numero_de_marcas_registradas",
    "ipqwery_num_patent_granted": "numero_de_patentes_otorgadas",
    "operating_status": "estado_operativo",
    "category": "categoria",
    "company_age": "antiguedad_de_la_empresa",
    "linkedin_employees": "empleados_en_linkedin",
    "linkedin_followers": "seguidores_en_linkedin",
    "facebook_followers": "seguidores_en_facebook",
    "twitter_followers": "seguidores_en_twitter",
    "num_articles": "numero_de_articulos",
    
    # Datos de fundadores
    "num_founders": "numero_total_de_fundadores",
    "num_female_founders": "numero_de_mujeres_fundadores",
    "num_male_founders": "numero_de_hombres_fundadores",
    
    # Informacion de fondeo
    "equity_funding_total": "financiacion_total_por_acciones",
    "last_equity_funding_total": "ultima_financiacion_por_acciones",
    "num_investments_funding_rounds": "numero_de_inversiones_en_rondas_de_financiacion",
    "num_investors": "numero_de_inversionistas",
    "num_lead_investors": "numero_de_inversionistas_principales",
    "funding_total": "financiacion_total",
    "num_funding_rounds": "numero_de_rondas_de_financiacion",
    "last_funding_type": "tipo_de_ultima_financiacion",
    "last_funding_at": "fecha_de_ultima_financiacion",
    "last_funding_total": "monto_de_ultima_financiacion",
     "years_since_last_funding": "años_desde_ultima_financiacion"
    
}

# Orden de columnas segun categorias
ordered_columns = [
     # Empresa
    "company_name", "founded_on", "ipqwery_num_trademark_registered",
    "ipqwery_num_patent_granted", "operating_status", "category", "company_age",
    "linkedin_employees", "linkedin_followers", "facebook_followers",
    "twitter_followers", "num_articles",
    
    # Fundadores
    "num_founders", "num_female_founders", "num_male_founders",
    
    # Fondeo
    "equity_funding_total", "last_equity_funding_total", "num_investments_funding_rounds",
    "num_investors", "num_lead_investors", "funding_total", "num_funding_rounds",
    "last_funding_type", "last_funding_at", "last_funding_total", "years_since_last_funding"
   
]

# Reordenar columnas y traducir nombres
df = df[ordered_columns].rename(columns=column_translation)

# Se filtra empresa que se encontró como outlier en el modelado
df_filtrado  = df[df['nombre_de_la_empresa'] != "Explo"].copy()

# Guardar 
df_filtrado .to_csv("datasets/startups_para_modelar.csv", index=False)
len(df_filtrado )

#df .to_csv("datasets/startups_para_modelar.csv", index=False)
#len(df )
